split tests_modeling_utils
